+++ crates/guest-write-file/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## [0.1.76](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.75...guest-write-file-v0.1.76) (2026-08-18)
+
 ## [0.1.75](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.74...guest-write-file-v0.1.75) (2026-08-17)
 
 ## [0.1.74](https://github.com/vm0-ai/vm0/compare/guest-write-file-v0.1.73...guest-write-file-v0.1.74) (2026-08-16)